LayoutTests/http/tests/site-isolation/history/add-iframes-and-navigate-mainframe.html

<!-- webkit-test-runner [ SiteIsolationEnabled=true ] -->
<script src="/js-test-resources/js-test.js"></script>
<script>
description("Verifies that navigating the main frame while adding an iframe only increments history.length by one.");
jsTestIsAsync = true;

onmessage = () => {
    setTimeout(() => {
        sessionStorage.didNavigate = true;
        location.href = 'data:text/html,<script> history.back(); <\/script>';
    }, 10);
}

onload = () => {
    if (sessionStorage.didNavigate) {
        delete sessionStorage.didNavigate;
        shouldBe("history.length", "3");
        finishJSTest();
    } else { 
        setInterval(() => {
            const iframe = document.createElement('iframe');
            iframe.src = 'http://localhost:8000/site-isolation/resources/green-background.html';
            document.body.appendChild(iframe);
        }, 1);
    }
}
</script>
